# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Enterprise Donation Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        mobile:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
        isActive:
          type: boolean
        isSuspended:
          type: boolean
        photoUrl:
          type: ["string", "null"]
        aadharNumber:
          type: ["string", "null"]
        fatherHusbandName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - mobile
        - role
        - isActive
        - createdAt
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        mobile:
          type: string
        role:
          type: string
          enum: [super_admin, admin, collector, public]
        password:
          type: string
        email:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        aadharNumber:
          type: ["string", "null"]
        fatherHusbandName:
          type: ["string", "null"]
      required:
        - name
        - mobile
        - role
    UpdateUserBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
        isSuspended:
          type: ["boolean", "null"]
        password:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        aadharNumber:
          type: ["string", "null"]
        fatherHusbandName:
          type: ["string", "null"]
...